small change to formatAddress function, added another test for that and moved test files into a folder together
